move float_operations to cuda/utils folder

+++ checker/pv/include/PrimaryVertexChecker.h
@@ -15,7 +15,7 @@
 
 // configuration for PV checker -> check values
 static constexpr int nTracksToBeRecble = 4;
-static constexpr double dzIsolated = 10.; // mm
+static constexpr float dzIsolated = 10.f; // mm
 static constexpr bool matchByTracks = false;
 
 void checkPVs(
